feat: separate some Notification shared files.
